Convert to Maven project
Also added a .gitignore to cover Maven compiled stuff
