+++ src/main/resources/net/TheDgtl/Stargate/resources/fr.txt
