+++ src/main/resources/net/TheDgtl/Stargate/resources/it.txt
