+++ src/main/resources/net/TheDgtl/Stargate/resources/nl.txt
