+++ src/main/resources/net/TheDgtl/Stargate/resources/pt-br.txt
