--- src/net/TheDgtl/Stargate/RelativeBlockVector.java
@@ -1,45 +1,0 @@
-package net.TheDgtl.Stargate;
-
-/**
- * Stargate - A portal plugin for Bukkit
- * Copyright (C) 2011 Shaun (sturmeh)
- * Copyright (C) 2011 Dinnerbone
- * Copyright (C) 2011, 2012 Steven "Drakia" Scott <Contact@TheDgtl.net>
- * 
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- * 
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-
-public class RelativeBlockVector {
-	private int right = 0;
-	private int depth = 0;
-	private int distance = 0;
-
-	public RelativeBlockVector(int right, int depth, int distance) {
-		this.right = right;
-		this.depth = depth;
-		this.distance = distance;
-	}
-
-	public int getRight() {
-		return right;
-	}
-
-	public int getDepth() {
-		return depth;
-	}
-
-	public int getDistance() {
-		return distance;
-	}
-}
