--- src/net/TheDgtl/Stargate/resources/ru.txt
@@ -1,22 +1,0 @@
-﻿prefix=[Портал] 
-teleportMsg=Вы перемещены
-destroyMsg=Портал уничтожен
-invalidMsg=Недопустимый пункт назначения
-blockMsg=Пункт назначения заблокирован
-destEmpty=Список пунктов назначения пуст
-denyMsg=Доступ запрещен
-
-ecoDeduct=Вычтено %cost%
-ecoRefund=Возвращено %cost%
-ecoObtain=Получено %cost% с портала %portal%
-ecoInFunds=Недостаточно средств
-
-createMsg=Портал создан
-createNetDeny=У вас нет доступа к этой сети порталов
-createGateDeny=У вас нет доступа к размещению портала
-createPersonal=Создание портала с персональной сетью
-createNameLength=Имя слишком короткое или слишком длинное
-createExists=Портал с таким именем уже существует
-createFull=Эта сеть порталов заполнена
-createWorldDeny=У вас нет доступа к этому миру
-createConflict=Портал конфликтует с уже существующим порталом
