--- src/plugin.yml
@@ -1,29 +1,0 @@
-name: Stargate
-main: net.TheDgtl.Stargate.Stargate
-version: 0.7.7.1
-description: Stargate mod for Bukkit
-author: Drakia
-website: http://www.thedgtl.net
-commands:
-  sg:
-    description: Used to reload the plugin. Console use only.
-    usage: /<command> reload - Used to reload the plugin. Console use only.
-permissions:
-  stargate.use:
-    description: Allow use of all gates linking to any world in any network
-    default: true
-  stargate.create:
-    description: Allow creating gates on any network
-    default: op
-  stargate.destroy:
-    description: Allow destruction gates on any network
-    default: op
-  stargate.free:
-    description: Allow free use/creation/destruction of gates
-    default: op
-  stargate.option:
-    description: Allow use of all options
-    default: op
-  stargate.admin:
-    description: Allow all admin features (Hidden/Private only so far)
-    default: op
